KB source formatting
git-svn-id: svn://svn.liferay.com/repos/public/plugins/trunk@75744 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portlets/knowledge-base-portlet/docroot/WEB-INF/src/com/liferay/knowledgebase/service/impl/ArticleLocalServiceImpl.java
@@ -1045,7 +1045,7 @@ public class ArticleLocalServiceImpl extends ArticleLocalServiceBaseImpl {
 		assetEntryLocalService.deleteEntry(
 			Article.class.getName(), article.getClassPK());
 
-		if (!article.isApproved()) {
+		if (!article.isApproved() && !article.isFirstVersion()) {
 			assetEntryLocalService.deleteEntry(
 				Article.class.getName(), article.getResourcePrimKey());
 		}
@@ -1056,7 +1056,7 @@ public class ArticleLocalServiceImpl extends ArticleLocalServiceBaseImpl {
 
 		deleteAttachments(article, article.getClassPK());
 
-		if (!article.isApproved()) {
+		if (!article.isApproved() && !article.isFirstVersion()) {
 			deleteAttachments(article, article.getResourcePrimKey());
 		}
 	}
